re-arrange folders and files
