+++ node/bin/cli.js
@@ -1,5 +1,5 @@
 #!/usr/bin/env node
 
-const run = require('../src/env0-deploy-cli');
+const run = require('../src/main');
 
 run();
