+++ node/src/commands/cancel.js
@@ -1,0 +1,3 @@
+const setDeploymentStatusApproval = require('../utils/set-deployment-approval-status');
+
+module.exports = setDeploymentStatusApproval('cancel');
